+++ app/src/main/java/com/example/fit_i_trainer/ui/profile/modify/ProfileModifyPicActivity.kt
@@ -1,4 +1,4 @@
-package com.example.fit_i_trainer.ui.profile
+package com.example.fit_i_trainer.ui.profile.modify
 
 import android.Manifest
 import android.net.Uri
