+++ app/src/main/java/com/example/fit_i_trainer/ui/profile/modify/ProfileModifyServiceActivity.kt
@@ -1,4 +1,4 @@
-package com.example.fit_i_trainer.ui.profile
+package com.example.fit_i_trainer.ui.profile.modify
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,20 +7,21 @@ import android.text.TextWatcher
 import android.widget.Button
 import android.widget.EditText
 import android.widget.ImageButton
-import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import com.example.fit_i_trainer.R
+import com.example.fit_i_trainer.ui.profile.ProfileActivity
 
-class ProfileModifyMeActivity :AppCompatActivity(){
+class ProfileModifyServiceActivity :AppCompatActivity(){
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_profile_modify_me)
-        //객체생성
-        val edit : EditText = findViewById(R.id.tv_about_me_context)
+        setContentView(R.layout.activity_profile_modify_service)
+        //객체 생성
+        val edit : EditText=findViewById(R.id.tv_about_service_context)
         val done = findViewById<Button>(R.id.btn_done)
-        val countText = findViewById<TextView>(R.id.tv_char_num)
 
         done.isEnabled = false
+
         edit.addTextChangedListener(object:TextWatcher{
             var maxText =""
             override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
@@ -28,25 +29,16 @@ class ProfileModifyMeActivity :AppCompatActivity(){
             }
 
             override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
-//                if(edit.length()>500){
-//                    Toast.makeText(this@ProfileModifyMeActivity,"최대 500자까지 입력 가능합니다.",
-//                    Toast.LENGTH_SHORT).show()
-//                    edit.setText(maxText)
-//                    edit.setSelection(edit.length())
-//                    countText.setText("${edit.length()}")
-//                }
+
             }
 
             override fun afterTextChanged(p0: Editable?) {
                 done.isEnabled = true
             }
         })
-
         //완료버튼
+
         done.setOnClickListener{
-//            val intent1 = Intent(this,ProfileAboutMeActivity::class.java)
-//            intent1.putExtra("contextMe",edit.text.toString())
-//            startActivity(intent1)
             val intent = Intent(this, ProfileActivity::class.java)
             startActivity(intent)
             finish()
